requirements.txt
alembic==1.18.4
blinker==1.9.0
click==8.3.1
colorama==0.4.6
dynaconf==3.2.12
Flask==3.1.3
Flask-Migrate==4.1.0
flask-restful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
pydantic[email]==2.10.6
python-dotenv==1.2.2
SQLAlchemy==2.0.48
typing_extensions==4.15.0
Werkzeug==3.1.6
PyJWT==2.10.1
pytest==8.3.5
